change value to long long and remove the buil directory

+++ is_prime_prog/main.cpp
@@ -1,17 +1,17 @@
 #include <iostream>
 
-bool is_prime(int n) {
+bool is_prime(long long n) {
     if (n <= 1) return false;
     if (n <= 3) return true;
     if (n % 2 == 0 || n % 3 == 0) return false;
-    for (int i = 5; i * i <= n; i = i + 6)
+    for (long long i = 5; i * i <= n; i = i + 6)
         if (n % i == 0 || n % (i + 2) == 0)
             return false;
     return true;
 }
 
 int main() {
-    int a;
+    long long a;
     while (std::cin >> a) {
         std::cout << a << " is a prime: " << (is_prime(a) ? "True" : "False") << std::endl;
     }
